refactor: small updates to details page and overview tab
