+++ src/components/quickstarts/QuickstartOverview.js
@@ -3,6 +3,23 @@ import { css } from '@emotion/react';
 
 import { quickstart } from '../../types';
 import OverviewTile from './OverviewTile';
+import Markdown from '../Markdown';
+
+const allowedElements = [
+  'h1',
+  'h2',
+  'h3',
+  'ol',
+  'ul',
+  'li',
+  'p',
+  'blockquote',
+  'code',
+  'a',
+  'strong',
+  'em',
+  'hr',
+];
 
 const QuickstartOverview = ({ quickstart }) => {
   return (
@@ -45,6 +62,17 @@ const QuickstartOverview = ({ quickstart }) => {
           />
         ))}
       </div>
+      {quickstart.description && (
+        <Markdown
+          skipHtml
+          allowedElements={allowedElements}
+          css={css`
+            margin: 2em 0;
+          `}
+        >
+          {quickstart.description}
+        </Markdown>
+      )}
     </>
   );
 };
